refactor: add unit-createpost module
