+++ core-commonui/src/androidMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/core/commonui/di/Utils.kt
@@ -1,7 +1,6 @@
 package com.github.diegoberaldin.raccoonforlemmy.core.commonui.di
 
 import com.github.diegoberaldin.raccoonforlemmy.core.commonui.communitydetail.CommunityDetailMviModel
-import com.github.diegoberaldin.raccoonforlemmy.core.commonui.createpost.CreatePostMviModel
 import com.github.diegoberaldin.raccoonforlemmy.core.commonui.postdetail.PostDetailMviModel
 import com.github.diegoberaldin.raccoonforlemmy.core.commonui.userdetail.UserDetailMviModel
 import com.github.diegoberaldin.raccoonforlemmy.domain.lemmy.data.CommunityModel
@@ -39,11 +38,5 @@ actual fun getUserDetailViewModel(user: UserModel, otherInstance: String): UserD
         clazz = UserDetailMviModel::class.java,
         parameters = { parametersOf(user, otherInstance) },
     )
-    return res
-}
-
-actual fun getCreatePostViewModel(editedPostId: Int?): CreatePostMviModel {
-    val res: CreatePostMviModel by inject(clazz = CreatePostMviModel::class.java,
-        parameters = { parametersOf(editedPostId) })
     return res
 }
