+++ core-commonui/unit-createpost/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/createpost/di/Utils.kt
@@ -1,0 +1,7 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.createpost.di
+
+import com.github.diegoberaldin.raccoonforlemmy.unit.createpost.CreatePostMviModel
+
+expect fun getCreatePostViewModel(
+    editedPostId: Int?,
+): CreatePostMviModel
