+++ settings.gradle.kts
@@ -55,5 +55,6 @@ include(":core-commonui:unit-reportlist")
 include(":core-commonui:unit-saveditems")
 include(":core-commonui:unit-createreport")
 include(":core-commonui:unit-createcomment")
+include(":core-commonui:unit-createpost")
 include(":core-commonui:detailopener-api")
 include(":core-commonui:detailopener-impl")
